requirements.txt
netifaces==0.11.0
PyYAML==6.0.1
